services/orders/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-prometheus
prometheus_client
